# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List owner projects",
        "tags": ["projects"]
      },
      "post": {
        "operationId": "createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Create and activate a project",
        "tags": ["projects"]
      }
    },
    "/api/v1/projects/{projectId}": {
      "patch": {
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Update a project",
        "tags": ["projects"]
      }
    },
    "/api/v1/projects/{projectId}/activate": {
      "post": {
        "operationId": "activateProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Make a project active",
        "tags": ["projects"]
      }
    },
    "/api/v1/projects/{projectId}/archive": {
      "post": {
        "operationId": "archiveProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedProjectCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponseDto"
                }
              }
            }
          }
        },
        "summary": "Archive a project",
        "tags": ["projects"]
      }
    },
    "/api/v1/projects/{projectId}/issues": {
      "post": {
        "operationId": "createIssue",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a task in a project",
        "tags": ["issues"]
      }
    },
...
    "/api/v1/projects/{projectId}/sprints": {
      "get": {
        "operationId": "listSprints",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List project Sprints",
        "tags": ["sprints"]
      },
      "post": {
        "operationId": "createSprint",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSprintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a planned Sprint",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/issues": {
      "post": {
        "operationId": "addIssueToSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintIssueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Assign a task to a Sprint",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/issues/bulk": {
      "post": {
        "operationId": "bulkAddIssuesToSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkSprintIssuesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Assign multiple tasks to a Sprint atomically",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/issues/create": {
      "post": {
        "operationId": "createIssueInSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a task directly in a Sprint",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/issues/{issueId}": {
      "delete": {
        "operationId": "removeIssueFromSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a task from a Sprint",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/start": {
      "post": {
        "operationId": "startSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedSprintCommandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Start a planned Sprint",
        "tags": ["sprints"]
      }
    },
    "/api/v1/sprints/{sprintId}/complete": {
      "post": {
        "operationId": "completeSprint",
        "parameters": [
          {
            "name": "sprintId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteSprintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SprintResponseDto"
                }
              }
            }
          }
        },
        "summary": "Complete an active Sprint",
        "tags": ["sprints"]
      }
    },
...
  },
  "info": {
    "title": "My Kanban API",
    "description": "Private API for the personal Kanban and Scrum board",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ProjectSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": ["kanban", "scrum"]
          },
          "version": {
            "type": "number"
          },
          "doneRetentionDays": {
            "type": "number",
            "enum": [7, 14, 30]
          }
        },
        "required": ["id", "name", "color", "mode", "version", "doneRetentionDays"]
      },
      "ProjectListResponseDto": {
        "type": "object",
        "properties": {
          "activeProjectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectSummaryDto"
            }
          }
        },
        "required": ["activeProjectId", "projects"]
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "color": {
            "type": "string",
            "maxLength": 32,
            "default": "primary"
          },
          "mode": {
            "type": "string",
            "enum": ["kanban", "scrum"],
            "default": "kanban"
          }
        },
        "required": ["name"]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "color": {
            "type": "string",
            "maxLength": 32
          },
          "mode": {
            "type": "string",
            "enum": ["kanban", "scrum"]
          },
          "doneRetentionDays": {
            "type": "number",
            "enum": [7, 14, 30]
          }
        },
        "required": ["version"]
      },
      "VersionedProjectCommandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["version"]
      },
      "ChecklistItemInputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 300
          },
          "isCompleted": {
            "type": "boolean",
            "default": false
          }
        },
        "required": ["title"]
      },
      "CreateIssueDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "maxLength": 50000,
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": ["task", "story", "bug", "chore"],
            "default": "task"
          },
          "priority": {
            "type": "string",
            "enum": ["urgent", "high", "medium", "low", "none"],
            "default": "medium"
          },
          "labels": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storyPoints": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isBlocked": {
            "type": "boolean",
            "default": false
          },
          "blockedReason": {
            "type": "string",
            "maxLength": 500,
            "nullable": true
          },
          "checklist": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemInputDto"
            }
          },
          "beforeIssueId": {
            "type": "string",
            "format": "uuid"
          },
          "afterIssueId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["title", "columnId"]
      },
      "ChecklistItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "required": ["id", "title", "isCompleted"]
      },
      "IssueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "sprintId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "columnId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["task", "story", "bug", "chore"]
          },
          "priority": {
            "type": "string",
            "enum": ["urgent", "high", "medium", "low", "none"]
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storyPoints": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "blockedReason": {
            "type": "string",
            "nullable": true
          },
          "checklist": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistItemResponseDto"
            }
          },
          "checklistIncompleteCount": {
            "type": "number",
            "minimum": 0
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "sprintId",
          "columnId",
          "title",
          "description",
          "type",
          "priority",
          "labels",
          "storyPoints",
          "dueDate",
          "isBlocked",
          "blockedReason",
          "checklist",
          "checklistIncompleteCount",
          "completedAt",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "SprintResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "completed"]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "plannedPoints": {
            "type": "number",
            "minimum": 0
          },
          "plannedIssueCount": {
            "type": "number",
            "minimum": 0
          },
          "completedPoints": {
            "type": "number",
            "minimum": 0
          },
          "completedIssueCount": {
            "type": "number",
            "minimum": 0
          },
          "incompletePoints": {
            "type": "number",
            "minimum": 0
          },
          "incompleteIssueCount": {
            "type": "number",
            "minimum": 0
          },
          "issueCount": {
            "type": "number",
            "minimum": 0,
            "description": "Current number of tasks assigned to the sprint"
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "goal",
          "status",
          "startDate",
          "endDate",
          "plannedPoints",
          "plannedIssueCount",
          "completedPoints",
          "completedIssueCount",
          "incompletePoints",
          "incompleteIssueCount",
          "issueCount",
          "completedAt",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "SprintListResponseDto": {
        "type": "object",
        "properties": {
          "sprints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SprintResponseDto"
            }
          }
        },
        "required": ["sprints"]
      },
      "CreateSprintDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "goal": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-14"
          }
        },
        "required": ["name", "startDate", "endDate"]
      },
      "SprintIssueDto": {
        "type": "object",
        "properties": {
          "issueId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["issueId"]
      },
      "BulkSprintIssuesDto": {
        "type": "object",
        "properties": {
          "issueIds": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": ["issueIds"]
      },
      "VersionedSprintCommandDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": ["version"]
      },
      "CompleteSprintDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "incompleteDestination": {
            "description": "Use backlog or the UUID of another planned sprint in the same project",
            "oneOf": [
              {
                "enum": ["backlog"]
              },
              {
                "type": "string",
                "format": "uuid"
              }
            ]
          }
        },
        "required": ["version", "incompleteDestination"]
      },
...
    }
  }
}